Rename statetransitiontable.rb to state_transition_table.rb
