+++ lib/racc/state_transition_table.rb
